# src/swagger.json
{
	"components": {
...
		"schemas": {
			"BasicResponseModel": {
				"description": "기본 응답 모델",
				"properties": {
					"message": {
						"type": "string",
						"description": "오류 메시지"
					},
					"success": {
						"type": "boolean",
						"description": "성공 실패 여부, true: 성공, false: 실패"
					},
					"data": {
						"description": "처리 내용 반환"
					}
				},
				"required": [
					"message",
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlugInImageModel": {
				"properties": {
					"imageID": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"order_num": {
						"type": "number",
						"format": "double"
					},
					"github_url": {
						"type": "string"
					},
					"registerDatetime": {
						"type": "string"
					},
					"updateDatetime": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"github_url",
					"registerDatetime",
					"updateDatetime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RuntimeBlockModel": {
				"properties": {
					"blockID": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"enabled": {
						"type": "number",
						"format": "double"
					},
					"order_num": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"blockID",
					"name",
					"enabled",
					"order_num"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RuntimeModel": {
				"properties": {
					"blockRuntimeID": {
						"type": "number",
						"format": "double"
					},
					"blockID": {
						"type": "string"
					},
					"imageID": {
						"type": "number",
						"format": "double"
					},
					"order_num": {
						"type": "number",
						"format": "double"
					},
					"container_url": {
						"type": "string"
					},
					"container_port": {
						"type": "string"
					},
					"container_env": {
						"type": "string"
					}
				},
				"required": [
					"blockRuntimeID",
					"blockID",
					"imageID",
					"order_num",
					"container_url",
					"container_port",
					"container_env"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RuntimeControlModel": {
				"properties": {
					"blockRuntimeID": {
						"type": "number",
						"format": "double"
					},
					"container_name": {
						"type": "string"
					},
					"container_state": {
						"type": "string"
					},
					"image_url": {
						"type": "string"
					},
					"cpu": {
						"type": "string"
					},
					"ram": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"env": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"blockRuntimeID",
					"container_name",
					"container_state",
					"image_url",
					"cpu",
					"ram",
					"path"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"KakaoChatResModel": {
				"properties": {
					"version": {
						"type": "string"
					},
					"template": {
						"$ref": "#/components/schemas/Template"
					}
				},
				"required": [
					"version",
					"template"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "main_backend",
		"version": "1.0.0",
		"description": "Main process for backend",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/account": {
			"get": {
				"operationId": "GetInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "uid",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "Modify",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"Account"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": []
			}
		},
...
		"/plugin": {
			"get": {
				"operationId": "GetInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "설치된 플러그인의 세부 정보를 반환",
				"tags": [
					"Plugin"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [
					{
						"description": "페이지 번호",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "한 페이지에 렌더링 할 데이터 건 수",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "특정 플러그인 idx 에 해당하는 정보 반환, null 이면 리스트 반환",
						"in": "query",
						"name": "imageID",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "RegisterPlugin",
				"responses": {
					"200": {
						"description": "등록 성공 / 실패 여부",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "새로운 플러그인 설치",
				"tags": [
					"Plugin"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlugInImageModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "ModifyPlugin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "기존 설치된 플러그인 변경",
				"tags": [
					"Plugin"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlugInImageModel"
							}
						}
					}
				}
			}
		},
...
		"/runtimeBlock": {
			"get": {
				"operationId": "GetInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"RuntimeBlock"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "blockID",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Insert",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"RuntimeBlock"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RuntimeBlockModel"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"RuntimeBlock"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RuntimeBlockModel"
							}
						}
					}
				}
			}
		},
...
		"/runtime": {
			"get": {
				"operationId": "GetInfo",
				"responses": {
					"200": {
						"description": "특정 런타임 세부 정보 혹은 런타임 목록",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "런타임 내용 확인",
				"tags": [
					"Runtime"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [
					{
						"description": "페이지 번호",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "한 페이지에 렌더링 할 데이터 건 수",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "런타임 idx",
						"in": "query",
						"name": "blockRuntimeID",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "blockID",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/runtime/register": {
			"post": {
				"operationId": "RegisterNewRuntime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "런타임 등록",
				"tags": [
					"Runtime"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RuntimeModel"
							}
						}
					}
				}
			}
		},
		"/runtime/state": {
			"post": {
				"operationId": "ContainerStateControl",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "런타임 컨테이너 관리 API",
				"tags": [
					"Runtime"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RuntimeControlModel"
							}
						}
					}
				}
			}
		},
		"/runtime/kakaochat": {
			"post": {
				"operationId": "KakaoChatRuntime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/KakaoChatResModel"
								}
							}
						}
					}
				},
				"tags": [
					"Runtime"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/KakaoChatReqModel"
							}
						}
					}
				}
			}
		},
		"/runtime/execute/{rid}": {
			"post": {
				"operationId": "Execute",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "런타임 실행",
				"tags": [
					"Runtime"
				],
				"security": [],
				"parameters": [
					{
						"description": "실행할 런타임 idx",
						"in": "path",
						"name": "rid",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/runtime/modify": {
			"put": {
				"operationId": "ModifyRuntime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "런타임 수정",
				"tags": [
					"Runtime"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RuntimeModel"
							}
						}
					}
				}
			}
		},
		"/runtime/{blockRuntimeID}": {
			"delete": {
				"operationId": "DeleteRuntime",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BasicResponseModel"
								}
							}
						}
					}
				},
				"summary": "런타임 삭제",
				"tags": [
					"Runtime"
				],
				"security": [
					{
						"passport-cookie": []
					}
				],
				"parameters": [
					{
						"description": "삭제할 런타임 idx",
						"in": "path",
						"name": "blockRuntimeID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}